feat(perf_hooks): implement monitorEventLoopDelay for Node.js compatibility
- Add internal module with IntervalHistogram wrapper class
- Implement enable/disable/reset methods on histogram instances
- Add C++ bindings for event loop delay monitoring
- Integrate Zig EventLoopDelayMonitor with Timer system
- Handle percentile validation (range: > 0 && <= 100)
- Pass all Node.js compatibility tests
