+++ src/js/node/perf_hooks.ts
@@ -1,6 +1,5 @@
 // Hardcoded module "node:perf_hooks"
 const { throwNotImplemented } = require("internal/shared");
-const { validateObject, validateNumber, validateInteger } = require("internal/validators");
 
 const createFunctionThatMasqueradesAsUndefined = $newCppFunction(
   "ZigGlobalObject.cpp",
@@ -14,24 +13,8 @@ const cppCreateHistogram = $newCppFunction("JSNodePerformanceHooksHistogram.cpp"
   figures: number,
 ) => import("node:perf_hooks").RecordableHistogram;
 
-// Private C++ bindings for event loop delay monitoring
-const cppMonitorEventLoopDelay = $newCppFunction(
-  "JSNodePerformanceHooksHistogramPrototype.cpp",
-  "jsFunction_monitorEventLoopDelay",
-  1,
-) as (resolution: number) => import("node:perf_hooks").RecordableHistogram;
-
-const cppEnableEventLoopDelay = $newCppFunction(
-  "JSNodePerformanceHooksHistogramPrototype.cpp",
-  "jsFunction_enableEventLoopDelay",
-  2,
-) as (histogram: import("node:perf_hooks").RecordableHistogram, resolution: number) => void;
-
-const cppDisableEventLoopDelay = $newCppFunction(
-  "JSNodePerformanceHooksHistogramPrototype.cpp",
-  "jsFunction_disableEventLoopDelay",
-  1,
-) as (histogram: import("node:perf_hooks").RecordableHistogram) => void;
+// Import monitorEventLoopDelay implementation
+const monitorEventLoopDelayImpl = require("internal/perf_hooks/monitorEventLoopDelay");
 
 var {
   Performance,
@@ -139,69 +122,6 @@ class PerformanceResourceTiming {
 }
 $toClass(PerformanceResourceTiming, "PerformanceResourceTiming", PerformanceEntry);
 
-// IntervalHistogram wrapper class for event loop delay monitoring
-class IntervalHistogram {
-  #histogram: import("node:perf_hooks").RecordableHistogram;
-  #resolution: number;
-  #enabled: boolean = false;
-
-  constructor(resolution: number) {
-    this.#resolution = resolution;
-    this.#histogram = cppMonitorEventLoopDelay(resolution);
-  }
-
-  enable() {
-    if (!this.#enabled) {
-      cppEnableEventLoopDelay(this.#histogram, this.#resolution);
-      this.#enabled = true;
-      return true;
-    }
-    return false;
-  }
-
-  disable() {
-    if (this.#enabled) {
-      cppDisableEventLoopDelay(this.#histogram);
-      this.#enabled = false;
-      return true;
-    }
-    return false;
-  }
-
-  reset() {
-    this.#histogram.reset();
-  }
-
-  get min() {
-    return this.#histogram.min;
-  }
-
-  get max() {
-    return this.#histogram.max;
-  }
-
-  get mean() {
-    return this.#histogram.mean;
-  }
-
-  get stddev() {
-    return this.#histogram.stddev;
-  }
-
-  get exceeds() {
-    return this.#histogram.exceeds;
-  }
-
-  get percentiles() {
-    return this.#histogram.percentiles;
-  }
-
-  percentile(p: number) {
-    validateNumber(p, "percentile", 0, 100);
-    return this.#histogram.percentile(p);
-  }
-}
-
 export default {
   performance: {
     mark(_) {
@@ -262,17 +182,7 @@ export default {
   PerformanceObserverEntryList,
   PerformanceNodeTiming,
   monitorEventLoopDelay: function monitorEventLoopDelay(options?: { resolution?: number }) {
-    if (arguments.length > 0) {
-      validateObject(options, "options");
-    }
-
-    let resolution = 10;
-    if (options?.resolution !== undefined) {
-      validateInteger(options.resolution, "options.resolution", 1);
-      resolution = options.resolution;
-    }
-
-    return new IntervalHistogram(resolution);
+    return monitorEventLoopDelayImpl(options);
   },
   createHistogram: function createHistogram(options?: {
     lowest?: number | bigint;
